extract SignInPromptBox into component
