+++ src/pages/Watchlist.jsx
@@ -1,20 +1,14 @@
 import React, { useEffect } from 'react';
 import PageTemplate from '../components/PageTemplate/PageTemplate';
 import CustomHeading from '../components/CustomHeading';
-import {
-    Center,
-    Link,
-    Text,
-    useColorModeValue,
-    VStack,
-} from '@chakra-ui/react';
+import { Link, Text, useColorModeValue, VStack } from '@chakra-ui/react';
 import { useUser } from '../contexts/useUser';
 import CustomBox from '../components/CustomBox';
 import StockInfoTable from '../components/StockTable/StockInfoTable';
 import { useStockSymbol } from '../contexts/useStockInfo';
 import LoadingSpinner from '../components/LoadingSpinner';
-import { useLoginSignupPopup } from '../contexts/useLoginSignupPopup';
 import { Link as ReactRouterLink } from 'react-router-dom';
+import SignInPromptBox from '../components/SignInPromptBox';
 
 const Watchlist = () => {
     const { user, watchlist } = useUser();
@@ -25,8 +19,6 @@ const Watchlist = () => {
         setWatchlistStockInfo,
         isWatchlistDataLoading,
     } = useStockSymbol();
-    const { setMode, onOpen } = useLoginSignupPopup();
-    const blueColor = useColorModeValue('blue.light', 'blue.dark');
 
     useEffect(() => {
         const tickers = [];
@@ -38,49 +30,23 @@ const Watchlist = () => {
         sortStocks(watchlistStockInfo, key, direction, setWatchlistStockInfo);
     };
 
-    const EmptyWatchlistPrompt = () => (
+    const EmptyWatchlistPromptBox = () => (
         <CustomBox w="100%">
             <VStack m={8}>
                 <Text>Your watchlist is empty!</Text>
                 <Text>
                     Add stocks form the
-                    <Link as={ReactRouterLink} to="/" color={blueColor} mx={1}>
+                    <Link
+                        as={ReactRouterLink}
+                        to="/"
+                        color={useColorModeValue('blue.light', 'blue.dark')}
+                        mx={1}
+                    >
                         Home
                     </Link>
                     page.
                 </Text>
             </VStack>
-        </CustomBox>
-    );
-
-    const SignInPrompt = () => (
-        <CustomBox w="100%">
-            <Center m={8}>
-                <Link
-                    as="button"
-                    onClick={() => {
-                        setMode('login');
-                        onOpen();
-                    }}
-                    color={blueColor}
-                    mx={1}
-                >
-                    Log In
-                </Link>
-                or
-                <Link
-                    as="button"
-                    onClick={() => {
-                        setMode('signup');
-                        onOpen();
-                    }}
-                    color={blueColor}
-                    mx={1}
-                >
-                    Sign Up
-                </Link>
-                to use the watchlist!
-            </Center>
         </CustomBox>
     );
 
@@ -97,10 +63,10 @@ const Watchlist = () => {
                             handleSortClick={handleSortClick}
                         />
                     ) : (
-                        <EmptyWatchlistPrompt />
+                        <EmptyWatchlistPromptBox />
                     )
                 ) : (
-                    <SignInPrompt />
+                    <SignInPromptBox text="to use the Watchlist!" w="100%" />
                 )}
             </VStack>
         </PageTemplate>
